Modifies Client & tests
